Modify last test
The test is supposed to check if the function correctly detects
that the string "r3ace3car" is a palindrome even if it contains
unevenly spaced numbers. The test should expect the function to return
true, but it expects false; i deem this to be an error. Modify the
value from false to true.